moving the UI code to todo_widgets
