+++ flutter_mobx/example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx_example/counter/counter.dart';
-import 'package:flutter_mobx_example/todos/todos.dart';
+import 'package:flutter_mobx_example/todos/todo_widgets.dart';
 
 void main() => runApp(MyApp());
